Metrics finished
* Type of attributes working with currency has been changed from BigDecimal to long due to incompatibility with Gauges

+++ order-service/src/main/java/cz/muni/fi/pv217/barewbeer/entity/OrderItem.java
@@ -18,7 +18,7 @@ public class OrderItem {
 
     public long productId;
 
-    public BigDecimal pricePerUnit;
+    public long pricePerUnit;
 
     public int count;
